astropath/requirements.txt
# Required packages for astropath

# Users
numpy>=1.18.0
scipy>=1.4
astropy>=4.2
IPython>=7.2.0
pytest>=6.0.0
pandas>=0.25
photutils>=1.0.1
healpy>=1.14
astropy-healpix>=0.5
ligo.skymap>=0.5

